// src/lib/server/schemas/rsstream.ts
export const accountDataSchema = z.object({
  pub: z.string(),
  epub: z.string(),
  username: z.string(),
  name: z.string(),
  email: z.string(), // encrypted
  validate: z.string().nullable().optional(), // encrypted validation code
  ref: z.string().optional(),
  host: z.string(),
  feeds: z.number(),
  subscribed: z.number(),
  reset: z.string().optional(), // encrypted reset code
  expiry: z.number().optional(),
  prev: z.string().optional(), // previous public key after password reset
})
// ...
export const inviteCodeSchema = z.object({
  code: z.string(),
  owner: z.string(),
  key: z.string().optional(),
})
// ...
export type AccountData = z.infer<typeof accountDataSchema>
export type FeedData = z.infer<typeof feedDataSchema>
export type FeedItemData = z.infer<typeof feedItemDataSchema>
export type InviteCode = z.infer<typeof inviteCodeSchema>